uupdate

+++ backend/src/services/uploadImages.js
@@ -6,21 +6,23 @@ export const uploadImages = async (files, folder = "default") => {
     files.map(async (file) => {
       try {
         const res = await cloudinary.uploader.upload(file.path, {
-          folder,
+          folder: folder,
           resource_type: "image"
         });
 
-        // xoá file local an toàn
-        fs.unlink(file.path, (err) => {
-          if (err) {
-            console.error("Delete file error:", err);
-          }
-        });
+        if (file.path) {
+          fs.unlink(file.path, (err) => {
+            if (err) {
+              console.error("Delete file error:", err);
+            }
+          });
+        }
 
         return res.secure_url;
       } catch (err) {
-        // nếu upload fail vẫn xoá file
-        fs.unlink(file.path, () => { });
+        if (file.path) {
+          fs.unlink(file.path, () => { });
+        }
         throw err;
       }
     })
